\name{audiology}
\alias{audiology}
\docType{data}
\title{
Audiology
}
\description{
The audiology data is comprised of attributes collected from patient history and assessment. The 
challenge for this data is to distinguish 23 audiological conditions and the norm using the collected attributes.
}
\usage{data(audiology)}
\format{
  A data frame with 226 observations on 70 variables.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Audiology+(Standardized)
}
\references{
Bareiss, E. Ray, & Porter, Bruce (1987). Protos: An Exemplar-Based Learning Apprentice. In the Proceedings of the 4th International Workshop on Machine Learning, 12-23, Irvine, CA: Morgan Kaufmann. 
}
\examples{
data(audiology)
## maybe str(audiology) ; plot(audiology) ...
}
\keyword{datasets}
