\name{dermatology}
\alias{dermatology}
\docType{data}
\title{
Dermatology
}
\description{
The differential diagnosis of erythemato-squamous diseases is a real
problem in dermatology. They all share the clinical features of
erythema and scaling, with very little differences. The diseases in
this group are psoriasis, seboreic dermatitis, lichen planus,
pityriasis rosea, cronic dermatitis, and pityriasis rubra pilaris.
Usually a biopsy is necessary for the diagnosis but unfortunately
these diseases share many histopathological features as
well. Another difficulty for the differential diagnosis is that a
disease may show the features of another disease at the beginning
stage and may have the characteristic features at the following stages.
}
\usage{data(dermatology)}
\format{
  A data frame with 366 observations on 35 variables.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Dermatology
}
\references{
G. Demiroz, H. A. Govenir, and N. Ilter, Learning Differential Diagnosis of Eryhemato-Squamous Diseases using Voting Feature Intervals. Aritificial Intelligence in Medicine 
}
\examples{
data(dermatology)
## maybe str(dermatology) ; plot(dermatology) ...
}
\keyword{datasets}
