\name{soybean}
\alias{soybean}
\docType{data}
\title{
Soybean
}
\description{
The soybean data is used to diagnose diseases typical for the plant. Several attributes were
collected from the sampled plant itself as well as its enviroment. Each plant is labelled with
a diagnosis and the goal is to deduce the diagnosis from the collected attributes.

}
\usage{data(soybean)}
\format{
  A data frame with 307 observations on 36 variables.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Soybean+%28Large%29
}
\references{
Tan, M., & Eshelman, L. (1988). Using weighted networks to represent classification knowledge in noisy domains. Proceedings of the Fifth International Conference on Machine Learning (pp. 121-134). Ann Arbor, Michigan: Morgan Kaufmann.

Fisher,D.H. & Schlimmer,J.C. (1988). Concept Simplification and Predictive Accuracy. Proceedings of the Fifth International Conference on Machine Learning (pp. 22-28). Ann Arbor, Michigan: Morgan Kaufmann.
}

\examples{
data(soybean)
## maybe str(soybean) ; plot(soybean) ...
}
\keyword{datasets}
